+++ .storybook/config.js
@@ -3,7 +3,7 @@ import { configure } from "@storybook/react";
 const req = require.context(
   "../packages",
   true,
-  /^((?!node_modules).)*\.stories\.js$/
+  /^((?!node_modules).)*\.(stories|stories.web)\.js$/
 );
 
 const loadStories = () => req.keys().forEach(filename => req(filename));
